Fix -Werror=conversion failure in trimcr().
While here avoid a separate strlen call in read_repos(), using the
length returned by trimcr().  The check for a NULL buf is removed as it
was already tested earlier in the function.

+++ tools.h
@@ -158,7 +158,7 @@ extern uint8_t yesflag;
 extern uint8_t noflag;
 
 extern int charcount(char *, char);
-extern int trimcr(char *);
+extern size_t trimcr(char *);
 extern void free_list(char **);
 extern void do_log(const char *, const char *, ...);
 extern void trunc_str(char *, char, int);
